+++ tests/opengl/opengl_fbconfig_test.c
@@ -8,7 +8,7 @@ static const char* g_BoolsToSting[2] = {"False", "True"};
 PalBool openglFBConfigTest()
 {
     // initialize the video system and create a window
-    PalResult result = palInitVideo(nullptr, nullptr);
+    PalResult result = palInitVideo(nullptr, nullptr, nullptr);
     if (result != PAL_RESULT_SUCCESS) {
         logResult(result, "Failed to initialize video");
         return PAL_FALSE;