+++ tests/video/char_event_test.c
@@ -5,17 +5,17 @@
 PalBool charEventTest()
 {
     palLog(nullptr, "Press Escape or click close button to close Test");
-    
-    PalResult result;
-    PalWindow* window = nullptr;
-    PalWindowCreateInfo createInfo = {0};
 
-    // event driver
-    PalEventDriver* eventDriver = nullptr;
+    // fill the event driver create info
     PalEventDriverCreateInfo eventDriverCreateInfo = {0};
+    eventDriverCreateInfo.allocator = nullptr; // default allocator
+    eventDriverCreateInfo.callback = nullptr;  // no callback dispatch
+    eventDriverCreateInfo.queue = nullptr;     // default queue
+    eventDriverCreateInfo.userData = nullptr;  // null
 
     // create the event driver
-    result = palCreateEventDriver(&eventDriverCreateInfo, &eventDriver);
+    PalEventDriver* eventDriver = nullptr;
+    PalResult result = palCreateEventDriver(&eventDriverCreateInfo, &eventDriver);
     if (result != PAL_RESULT_SUCCESS) {
         logResult(result, "Failed to create event driver");
         return PAL_FALSE;
@@ -24,13 +24,14 @@ PalBool charEventTest()
     // initialize the video system. We pass the event driver to recieve video
     // related events the video system does not copy the event driver, it must
     // be valid till the video system is shutdown
-    result = palInitVideo(nullptr, eventDriver);
+    result = palInitVideo(nullptr, eventDriver, nullptr);
     if (result != PAL_RESULT_SUCCESS) {
         logResult(result, "Failed to initialize video");
         return PAL_FALSE;
     }
 
     // fill the create info struct
+    PalWindowCreateInfo createInfo = {0};
     createInfo.monitor = nullptr; // use default monitor
     createInfo.height = 480;
     createInfo.width = 640;
@@ -46,6 +47,7 @@ PalBool charEventTest()
     }
 
     // create the window with the create info struct
+    PalWindow* window = nullptr;
     result = palCreateWindow(&createInfo, &window);
     if (result != PAL_RESULT_SUCCESS) {
         logResult(result, "Failed to create window");