+++ tests/video/icon_test.c
@@ -5,26 +5,17 @@
 PalBool iconTest()
 {
     palLog(nullptr, "Press Escape or click close button to close Test");
-    
-    PalResult result;
-    PalWindow* window = nullptr;
-    PalIcon* icon = nullptr;
-    PalWindowCreateInfo createInfo = {0};
-    PalIconCreateInfo iconCreateInfo = {0};
-    PalBool running = PAL_FALSE;
-
-    // event driver
-    PalEventDriver* eventDriver = nullptr;
-    PalEventDriverCreateInfo eventDriverCreateInfo = {0};
 
     // fill the event driver create info
+    PalEventDriverCreateInfo eventDriverCreateInfo = {0};
     eventDriverCreateInfo.allocator = nullptr; // default allocator
     eventDriverCreateInfo.callback = nullptr;  // for callback dispatch
     eventDriverCreateInfo.queue = nullptr;     // default queue
     eventDriverCreateInfo.userData = nullptr;  // null
 
     // create the event driver
-    result = palCreateEventDriver(&eventDriverCreateInfo, &eventDriver);
+    PalEventDriver* eventDriver = nullptr;
+    PalResult result = palCreateEventDriver(&eventDriverCreateInfo, &eventDriver);
     if (result != PAL_RESULT_SUCCESS) {
         logResult(result, "Failed to create event driver");
         return PAL_FALSE;
@@ -33,7 +24,7 @@ PalBool iconTest()
     // initialize the video system. We pass the event driver to recieve video
     // related events the video does not copy this, this must be valid till the
     // video system is shutdown
-    result = palInitVideo(nullptr, eventDriver);
+    result = palInitVideo(nullptr, eventDriver, nullptr);
     if (result != PAL_RESULT_SUCCESS) {
         logResult(result, "Failed to initialize video");
         return PAL_FALSE;
@@ -71,10 +62,12 @@ PalBool iconTest()
     }
 
     // create icon
+    PalIconCreateInfo iconCreateInfo = {0};
     iconCreateInfo.width = 32;
     iconCreateInfo.height = 32;
     iconCreateInfo.pixels = pixels; // can be loaded from an .ico file
 
+    PalIcon* icon = nullptr;
     result = palCreateIcon(&iconCreateInfo, &icon);
     if (result != PAL_RESULT_SUCCESS) {
         logResult(result, "Failed to create window icon");
@@ -82,6 +75,7 @@ PalBool iconTest()
     }
 
     // fill the create info struct
+    PalWindowCreateInfo createInfo = {0};
     createInfo.monitor = nullptr; // use default monitor
     createInfo.height = 480;
     createInfo.width = 640;
@@ -90,6 +84,7 @@ PalBool iconTest()
     createInfo.title = "Icon Window";
 
     // create the window with the create info struct
+    PalWindow* window = nullptr;
     result = palCreateWindow(&createInfo, &window);
     if (result != PAL_RESULT_SUCCESS) {
         logResult(result, "Failed to create window");
@@ -97,9 +92,7 @@ PalBool iconTest()
     }
 
     // set the dispatch mode for window close event to recieve it
-    palSetEventDispatchMode(eventDriver, PAL_EVENT_WINDOW_CLOSE,
-                            PAL_DISPATCH_POLL); // polling
-
+    palSetEventDispatchMode(eventDriver, PAL_EVENT_WINDOW_CLOSE, PAL_DISPATCH_POLL);
     palSetEventDispatchMode(eventDriver, PAL_EVENT_KEYDOWN, PAL_DISPATCH_POLL);
 
     // set the icon
@@ -109,7 +102,7 @@ PalBool iconTest()
         return PAL_FALSE;
     }
 
-    running = PAL_TRUE;
+    PalBool running = PAL_TRUE;
     while (running) {
         // update the video system to push video events
         palUpdateVideo();