+++ tests/video/native_instance_test.c
@@ -310,14 +310,16 @@ PalBool nativeInstanceTest()
 {
     palLog(nullptr, "Press Escape or click close button to close Test");
     
-    PalResult result;
-
-    // event driver
-    PalEventDriver* eventDriver = nullptr;
+    // fill the event driver create info
     PalEventDriverCreateInfo eventDriverCreateInfo = {0};
+    eventDriverCreateInfo.allocator = nullptr; // default allocator
+    eventDriverCreateInfo.callback = nullptr;  // no callback dispatch
+    eventDriverCreateInfo.queue = nullptr;     // default queue
+    eventDriverCreateInfo.userData = nullptr;  // null
 
     // create the event driver
-    result = palCreateEventDriver(&eventDriverCreateInfo, &eventDriver);
+    PalEventDriver* eventDriver = nullptr;
+    PalResult result = palCreateEventDriver(&eventDriverCreateInfo, &eventDriver);
     if (result != PAL_RESULT_SUCCESS) {
         logResult(result, "Failed to create event driver");
         return PAL_FALSE;
@@ -330,15 +332,11 @@ PalBool nativeInstanceTest()
         return PAL_FALSE;
     }
 
-    // tell the video system to use out instance rather
-    // than creating a new one
-    // this can be set to the opengl system as well
-    palSetPreferredInstance(instance);
-
     // initialize the video system. We pass the event driver to recieve video
     // related events the video system does not copy the event driver, it must
     // be valid till the video system is shutdown
-    result = palInitVideo(nullptr, eventDriver);
+    // tell the video system to use out instance rather than creating a new one
+    result = palInitVideo(nullptr, eventDriver, instance);
     if (result != PAL_RESULT_SUCCESS) {
         logResult(result, "Failed to initialize video");
         return PAL_FALSE;