+++ tests/video/window_test.c
@@ -145,24 +145,17 @@ static void PAL_CALL onEvent(
 PalBool windowTest()
 {
     palLog(nullptr, "Press Escape or click close button to close Test");
-    
-    PalResult result;
-    PalWindow* window = nullptr;
-    PalWindowCreateInfo createInfo = {0};
-    PalBool running = PAL_FALSE;
-
-    // event driver
-    PalEventDriver* eventDriver = nullptr;
-    PalEventDriverCreateInfo eventDriverCreateInfo = {0};
 
     // fill the event driver create info
+    PalEventDriverCreateInfo eventDriverCreateInfo = {0};
     eventDriverCreateInfo.allocator = nullptr; // default allocator
     eventDriverCreateInfo.callback = onEvent;  // for callback dispatch
     eventDriverCreateInfo.queue = nullptr;     // default queue
     eventDriverCreateInfo.userData = nullptr;  // null
 
     // create the event driver
-    result = palCreateEventDriver(&eventDriverCreateInfo, &eventDriver);
+    PalEventDriver* eventDriver = nullptr;
+    PalResult result = palCreateEventDriver(&eventDriverCreateInfo, &eventDriver);
     if (result != PAL_RESULT_SUCCESS) {
         logResult(result, "Failed to create event driver");
         return PAL_FALSE;
@@ -193,13 +186,14 @@ PalBool windowTest()
     // initialize the video system. We pass the event driver to recieve video
     // related events the video system does not copy the event driver, it must
     // be valid till the video system is shutdown
-    result = palInitVideo(nullptr, eventDriver);
+    result = palInitVideo(nullptr, eventDriver, nullptr);
     if (result != PAL_RESULT_SUCCESS) {
         logResult(result, "Failed to initialize video");
         return PAL_FALSE;
     }
 
     // fill the create info struct
+    PalWindowCreateInfo createInfo = {0};
     createInfo.monitor = nullptr; // use default monitor
     createInfo.height = 480;
     createInfo.width = 640;
@@ -251,6 +245,7 @@ PalBool windowTest()
 #endif // NO_MAXIMIZEBOX
 
     // create the window with the create info struct
+    PalWindow* window = nullptr;
     result = palCreateWindow(&createInfo, &window);
     if (result != PAL_RESULT_SUCCESS) {
         logResult(result, "Failed to create window");
@@ -268,7 +263,7 @@ PalBool windowTest()
     }
 #endif // MAKE_TRANSPARENT
 
-    running = PAL_TRUE;
+    PalBool running = PAL_TRUE;
     while (running) {
         // update the video system to push video events
         palUpdateVideo();