+++ CHANGELOG.md
@@ -156,6 +156,7 @@ palJoinThread(thread, &retval);
 - **Video:** **palGetWindowHandleInfo()** now takes an info parameter.
 - **Video:** **palGetMouseDelta()** now takes floats instead of uint32_t.
 - **Video:** **palGetMouseWheelDelta()** now takes floats instead of uint32_t.
+- **Video:** **palInitVideo()** now takes a preferredInstance parameter.
 
 ### Removed
 - **Core:** Removed **PAL_RESULT_NULL_POINTER** result code.
@@ -184,7 +185,7 @@ palJoinThread(thread, &retval);
 - **Video:** Removed **palGetWindowHandleInfoEx** function.
 - **Video:** Removed **palGetRawMouseWheelDelta** function.
 
-
+- **Video:** Removed **palSetPreferredInstance** function.
 
 ### Tests